// src/generated/models/seller-create.ts
/* tslint:disable */

/**
 * BFF Service
 * Backend For Frontend service aggregating microservices
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: you@example.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 *
 * @export
 * @interface SellerCreate
 */
export interface SellerCreate {
  /**
   *
   * @type {string}
   * @memberof SellerCreate
   */
  name: string
  /**
   *
   * @type {string}
   * @memberof SellerCreate
   */
  email: string
  /**
   *
   * @type {string}
   * @memberof SellerCreate
   */
  phone: string
  /**
   *
   * @type {string}
   * @memberof SellerCreate
   */
  city: string
  /**
   *
   * @type {string}
   * @memberof SellerCreate
   */
  country: string
}